refactor(cli): split generic parser and help modules
Keep the extracted ccusage-cli crate from becoming another large ccusage-specific lib.rs. Move generic argument tokenization into arg_parser.rs, generic help rendering into help.rs, and build-time help source generation into a tested help_codegen module.

Move ccusage command/help shape into cli-commands.json so help.rs only routes generic command pages and option sets. The build script now pre-renders combined option strings and generated help page data from JSON at build time, keeping runtime dependency-free and lookup-based.

Also address the review finding where the deprecated global -a agent-filter check rejected the valid blocks active flag. Exact -a is now allowed for ccusage blocks and ccusage claude blocks, with parser regressions for both forms.

Validation: direnv exec . cargo test --manifest-path rust/Cargo.toml -p ccusage-cli; direnv exec . env -u CFLAGS cargo check --manifest-path rust/Cargo.toml -p ccusage-cli -p ccusage; direnv exec . env -u CFLAGS pnpm run test; direnv exec . pnpm run format.
